refactor custom errors and added input validations to controllers

+++ src/controllers/productsController.js
@@ -1,5 +1,6 @@
 import { productsRepository } from '../repository/productsRepository.js'
-import { customError } from '../models/errors/errorsDictionary.js'
+import { CustomError } from '../models/errors/customError.js'
+import { winstonLogger as logger } from '../utils/winstonLogger.js'
 
 
 export async function handleGet(req, res, next) {
@@ -8,7 +9,7 @@ export async function handleGet(req, res, next) {
             //---Productos---
             const{query,limit,page,sort} = req.query//DTO
             req.result = await productsRepository.getProducts(query,limit,page,sort)
-            //console.log(JSON.stringify(req.result))
+            
             req.statusCode = req.result? 200 : 404
             next()
         }
@@ -20,24 +21,29 @@ export async function handleGet(req, res, next) {
         } 
     }
     catch(error){
-        //res.status(req.statusCode).send({status:"failure", payload:"not found"})
         next(error)
     }
 }
 
 export async function handlePost(req, res, next) {
     try{
+        if(!req.body){
+            throw new CustomError("!req.body", 400)
+        }
         req.result = await productsRepository.createProduct(req.body)//DTO?
         req.statusCode = req.result? 201 : 400
         next()
     }
     catch(error){
-        res.status(req.statusCode).send({status:"failure", payload:"not found"})
+        next(error)
     }
 }
 
 export async function handlePostAndGetAll(req, res, next) {
     try{
+        if(!req.body){
+            throw new CustomError("!req.body", 400)
+        }
         req.result = await productsRepository.createProduct(req.body)
         req.statusCode = req.result? 201 : 400
         const{query,limit,page,sort} = req.query
@@ -46,29 +52,38 @@ export async function handlePostAndGetAll(req, res, next) {
         next()
     }
     catch(error){
-        res.status(500).send({status:"failure", payload:"not found"})
+        next(error)
     }
 }
 
 export async function handlePut(req, res, next) {
     try{
+        if(!req.body){
+            throw new CustomError("!req.body", 400)
+        }
+        if(!req.params._id){
+            throw new CustomError("!req.params._id", 400)
+        }
         req.result = await productsRepository.updateProduct(req.params._id,req.body)
         req.statusCode = req.result? 200 : 412
         next()
     }
     catch(error){
-        res.status(req.statusCode).send({status:"failure", payload:"not found"}) //TO DO: reemplazar los strings segun el codigo de erro
+        next(error)
     }
 }
 
 export async function handleDelete(req, res, next) {
     try{
+        if(!req.params._id){
+            throw new CustomError("!req.params._id", 400)
+        }
         req.result = await productsRepository.deleteProduct(req.params._id)
         req.statusCode = req.result? 200 : 404
         next()
     }
-    catch(err){
-        res.status(req.statusCode).send({status:"failure", payload:"not found"})
+    catch(error){
+        next(error)
     }
 }
 
@@ -83,7 +98,7 @@ export async function handleDeleteAndGetAll(req, res, next) {
         next()
     }
     catch(error){
-        res.status(500).send({status:"failure", payload:"not found"})
+        next(error)
     }
 }
 //TESTS
